etl_service/requirements.txt
elasticsearch>=8.11.0,<9
psycopg2-binary>=2.9.9
